refactor: cleaning and rearrange code structure
