+++ app/components/QueryProvider.tsx
@@ -1,5 +1,6 @@
 "use client";
 import React from "react";
+
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 const QueryProvider = ({ children }: { children: React.ReactNode }) => {
