+++ app/components/atom/index.ts
@@ -1,0 +1,13 @@
+import AlertAnimation from './AlertAnimation';
+import BackButton from './BackButton';
+import Button from './Button';
+import Counter from './Counter';
+import Input from './GeneralInput'
+import ImagesProduct from './ImagesProduct';
+import LoadingAnimation from './LoadingAnimation';
+import SearchBar from './SearchBar';
+import Selector from './Selector';
+import ToggleVisiblePassword from './ToggleVisiblePassword';
+import UserAnimation from './UserAnimation';
+
+export { AlertAnimation, BackButton, Button, Counter, Input, ImagesProduct, LoadingAnimation, SearchBar, Selector, ToggleVisiblePassword, UserAnimation }
