+++ app/components/index.ts
@@ -1,18 +1,7 @@
-export { default as Layout } from "./Layout"
-export { default as Header } from "./Header"
-export { default as HamburgerMenu } from "./HamburgerMenu"
-export { default as HeroBanner } from "./HeroBanner"
-export { default as Button } from './Button'
-export { default as Footer } from './Footer'
-export { default as Products } from "./ProductListWrapper"
-export { default as FormUser } from "./FormUser"
-export { default as UserAnimation } from "./UserAnimation"
-export { default as CartProduct } from './CartProduct'
-export { default as ImagesProduct } from './ImagesProduct'
-export { default as ImageSlider } from './ImageSlider'
-export { default as ErrorRelogin } from './ErrorRelogin'
-export { default as CartProductLoading } from './CartProductLoading'
-export { default as Modals } from './Modals'
-export { default as BackButton } from './BackButton'
-export { default as Input } from './GeneralInput'
-export { default as Dropzone } from './Dropzone'
+export { default as Layout } from "./templates/Layout"
+export { default as ErrorRelogin } from './templates/ErrorRelogin'
+
+export * from './atom'
+export * from './molecule'
+export * from './organism'
+export * from './templates'
