+++ app/components/molecule/Dropzone.tsx
@@ -1,9 +1,10 @@
 "use client";
-import Image from "next/image";
 import React from "react";
+import Image from "next/image";
+
 import { useDropzone } from "react-dropzone";
 import useWindowsDimension from "@/utils/setWidth";
-import Button from "./Button";
+import { Button } from "../atom";
 import ImageSlider from "./ImageSlider";
 
 interface DropzoneProps {
