+++ app/components/molecule/HeroBanner.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import Link from "next/link";
-import Button from "./Button";
 import Image from "next/image";
+
+import { Button } from "../atom";
 
 const HeroBanner: React.FC = () => {
   return (
