+++ app/components/molecule/ImageSlider.tsx
@@ -1,6 +1,7 @@
 "use client";
-import Image from "next/image";
 import React from "react";
+import Image from "next/image";
+
 import { Swiper, SwiperSlide } from "swiper/react";
 
 interface ImageSliderProps {
