+++ app/components/organism/index.ts
@@ -1,0 +1,5 @@
+import LoginButton from './LoginButton'
+import Products from './ProductListWrapper'
+import ProductList from './ProductsList'
+
+export { LoginButton, Products, ProductList }
