+++ app/components/templates/Layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "./Header";
-import Footer from "./Footer";
+
+import { Header, Footer } from "../molecule";
 
 interface LayoutProps {
   children: React.ReactNode;
