+++ app/components/templates/index.ts
@@ -1,0 +1,4 @@
+import ErrorRelogin from './ErrorRelogin';
+import Layout from './Layout';
+
+export { ErrorRelogin, Layout }
