+++ app/user/[id]/products/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const page = () => {
+  return <div>page</div>;
+};
+
+export default page;
