+++ utils/setWidth.ts
@@ -1,29 +1,5 @@
 import { useState, useEffect } from "react";
 
-// function getWindowDimensions() {
-//   if (typeof window !== 'undefined') {
-//     const { innerWidth: width, innerHeight: height } = window;
-//     return { width, height };
-//   }
-// }
-
-// export default function useWindowsDimension() {
-//   const [windowsDimension, setWindowsDimension] = useState(
-//     getWindowDimensions()
-//   );
-
-//   useEffect(() => {
-//     function handleResize() {
-//       setWindowsDimension(getWindowDimensions());
-//     }
-
-//     window.addEventListener("resize", handleResize);
-//     return () => window.removeEventListener("resize", handleResize);
-//   }, []);
-
-//   return windowsDimension;
-// }
-
 export default function useWindowsDimension() {
   const [windowsDimension, setWindowsDimension] = useState(
     { width: 0, height: 0 }
